requirement.txt
streamlit == 1.46.1
Pillow == 11.2.1
torch==2.9.0
torchvision==0.24.0